refactor: reorganize project structure (FSD)
