+++ public/Svgs/Icons/SvgIndex.tsx
@@ -1,76 +1,151 @@
-import React from "react";
+import React from 'react';
 export const ICON_DATA = {
-  // '전체' 아이콘
-  AllIcon: (isActive: boolean) => (
-    <svg xmlns="http://www.w3.org/2000/svg" width="14" height="12" viewBox="0 0 12 12" fill="none">
-  <path d="M8 12C7.63333 12 7.31956 11.8696 7.05867 11.6087C6.79778 11.3478 6.66711 11.0338 6.66667 10.6667V8C6.66667 7.63333 6.79733 7.31956 7.05867 7.05867C7.32 6.79778 7.63378 6.66711 8 6.66667H10.6667C11.0333 6.66667 11.3473 6.79733 11.6087 7.05867C11.87 7.32 12.0004 7.63378 12 8V10.6667C12 11.0333 11.8696 11.3473 11.6087 11.6087C11.3478 11.87 11.0338 12.0004 10.6667 12H8ZM8 5.33333C7.63333 5.33333 7.31956 5.20289 7.05867 4.942C6.79778 4.68111 6.66711 4.36711 6.66667 4V1.33333C6.66667 0.966667 6.79733 0.652889 7.05867 0.392C7.32 0.131111 7.63378 0.000444444 8 0H10.6667C11.0333 0 11.3473 0.130667 11.6087 0.392C11.87 0.653333 12.0004 0.967111 12 1.33333V4C12 4.36667 11.8696 4.68067 11.6087 4.942C11.3478 5.20333 11.0338 5.33378 10.6667 5.33333H8ZM1.33333 5.33333C0.966667 5.33333 0.652889 5.20289 0.392 4.942C0.131111 4.68111 0.000444444 4.36711 0 4V1.33333C0 0.966667 0.130667 0.652889 0.392 0.392C0.653333 0.131111 0.967111 0.000444444 1.33333 0H4C4.36667 0 4.68067 0.130667 4.942 0.392C5.20333 0.653333 5.33378 0.967111 5.33333 1.33333V4C5.33333 4.36667 5.20289 4.68067 4.942 4.942C4.68111 5.20333 4.36711 5.33378 4 5.33333H1.33333ZM1.33333 12C0.966667 12 0.652889 11.8696 0.392 11.6087C0.131111 11.3478 0.000444444 11.0338 0 10.6667V8C0 7.63333 0.130667 7.31956 0.392 7.05867C0.653333 6.79778 0.967111 6.66711 1.33333 6.66667H4C4.36667 6.66667 4.68067 6.79733 4.942 7.05867C5.20333 7.32 5.33378 7.63378 5.33333 8V10.6667C5.33333 11.0333 5.20289 11.3473 4.942 11.6087C4.68111 11.87 4.36711 12.0004 4 12H1.33333Z" fill={isActive ? "#FFFFFF" : "#596373"}/>
-</svg>
-  ),
-  // '상의' 아이콘
-  TopIcon: (isActive: boolean) => (
-    <svg xmlns="http://www.w3.org/2000/svg" width="14" height="12" viewBox="0 0 14 12" fill="none">
-  <path d="M8.58867 0.0048273L8.652 0.000160658L8.72667 0.00282729L8.802 0.0141606L8.87733 0.0341607L12.8773 1.36749C12.9976 1.40758 13.1039 1.48111 13.1838 1.57947C13.2638 1.67783 13.314 1.79693 13.3287 1.92283L13.3333 2.00016V5.33349C13.3333 5.49678 13.2734 5.65439 13.1649 5.77641C13.0563 5.89843 12.9068 5.97639 12.7447 5.99549L12.6667 6.00016H11.3333V10.6668C11.3334 11.0032 11.2064 11.3272 10.9777 11.5739C10.7489 11.8205 10.4354 11.9716 10.1 11.9968L10 12.0002H3.33333C2.99695 12.0003 2.67296 11.8732 2.4263 11.6445C2.17965 11.4158 2.02856 11.1023 2.00333 10.7668L2 10.6668V6.00016H0.666667C0.503378 6.00014 0.345775 5.94019 0.223752 5.83168C0.101729 5.72318 0.0237715 5.57366 0.00466668 5.41149L3.80479e-08 5.33349V2.00016C-4.27825e-05 1.87339 0.0360591 1.74924 0.104072 1.64226C0.172085 1.53528 0.26919 1.44991 0.384 1.39616L0.456 1.36749L4.456 0.0341607C4.55622 0.000780185 4.66294 -0.00830679 4.76736 0.00764891C4.87179 0.0236046 4.97092 0.0641461 5.0566 0.125932C5.14228 0.187719 5.21206 0.268981 5.26017 0.363022C5.30828 0.457064 5.33336 0.561193 5.33333 0.666827C5.33178 1.01282 5.46479 1.34586 5.70425 1.5956C5.94371 1.84533 6.27088 1.99221 6.61663 2.00519C6.96237 2.01817 7.29963 1.89624 7.55714 1.66517C7.81466 1.43409 7.97227 1.11196 7.99667 0.766827L8.00267 0.606827L8.014 0.531494L8.03867 0.442161L8.068 0.373494L8.10133 0.312161L8.14667 0.250161L8.19267 0.196827C8.22956 0.161272 8.26844 0.130161 8.30933 0.103494L8.37333 0.0681607L8.442 0.0388274L8.514 0.017494L8.58867 0.0048273Z" fill={isActive ? "#FFFFFF" : "#596373"}/>
-</svg>
-  ),
-   BottomIcon: (isActive: boolean) => (
-     <svg xmlns="http://www.w3.org/2000/svg" width="11" height="15" viewBox="0 0 11 15" fill="none">
-  <path d="M5.16506 9.62867e-05C3.77269 0.00594004 2.43881 0.230596 1.46 0.656346C0.9545 5.26441 0.43525 9.99469 0 13.961L3.8545 14.7365L4.92191 6.5226H5.56659L6.63691 14.7648H6.64372L10.6769 13.9435C10.186 9.51497 9.70747 5.085 9.21681 0.656346C8.00753 0.209252 6.5575 -0.00521621 5.16509 9.62867e-05H5.16506Z" fill={isActive ? "#FFFFFF" : "#596373"}/>
-</svg>
-  ),
-   OuterIcon: (isActive: boolean) => (
-    <svg xmlns="http://www.w3.org/2000/svg" width="15" height="14" viewBox="0 0 15 14" fill="none">
-  <path fill-rule="evenodd" clip-rule="evenodd" d="M8.49637 0C8.84578 2.37364e-05 9.19038 0.0814238 9.50287 0.23775L11.0254 0.99975C11.4829 1.22851 11.8888 1.54846 12.2181 1.93991C12.5474 2.33137 12.7931 2.78605 12.9401 3.276L14.6366 8.93025C14.7382 9.26864 14.7171 9.63203 14.577 9.95638C14.437 10.2807 14.1869 10.5452 13.8709 10.7032L13.1119 11.082C12.9648 11.1556 12.7976 11.1787 12.6361 11.148C12.4746 11.1172 12.3276 11.0342 12.2179 10.9117L11.8504 10.5L11.6494 11.814C11.5243 12.2969 11.2424 12.7245 10.8479 13.0297C10.4534 13.3348 9.96865 13.5003 9.46987 13.5H5.23087C4.73209 13.5003 4.24735 13.3348 3.85284 13.0297C3.45832 12.7245 3.1764 12.2969 3.05137 11.814L2.85037 10.5L2.48287 10.9117C2.37308 11.0344 2.22599 11.1175 2.0643 11.1482C1.9026 11.179 1.73528 11.1557 1.58812 11.082L0.829121 10.7032C0.513116 10.5452 0.263032 10.2807 0.122946 9.95638C-0.0171411 9.63203 -0.0382387 9.26864 0.0633712 8.93025L1.75987 3.276C1.9069 2.78605 2.15262 2.33137 2.48191 1.93991C2.81119 1.54846 3.21709 1.22851 3.67462 0.99975L5.19787 0.23775C5.51036 0.0814238 5.85496 2.37364e-05 6.20437 0H8.49637ZM8.49637 1.5H6.20437C6.1174 1.5001 6.03112 1.51533 5.94937 1.545L5.86912 1.5795L5.14912 1.9395L7.35037 5.3625L9.55162 1.938L8.83162 1.578C8.75376 1.53929 8.66976 1.51442 8.58337 1.5045L8.49637 1.5Z" fill={isActive ? "#FFFFFF" : "#596373"}/>
-</svg>
-  ),
-   DressIcon: (isActive: boolean) => (
-     <svg xmlns="http://www.w3.org/2000/svg" width="13" height="14" viewBox="0 0 13 14" fill="none">
-  <path d="M4.12812 0L3.49375 0.318125C3.49375 1.31906 3.51875 3.06375 3.92812 4.41062C5.75 4.85437 6.66875 4.85437 8.49063 4.41062C8.90625 3.06687 8.925 1.70625 8.92813 0.318125L8.29063 0C8.15313 0.179375 7.925 0.47 7.60937 0.830312C7.38437 1.08531 7.14688 1.34156 6.91875 1.54281C6.80625 1.64344 6.69688 1.73031 6.5875 1.79813C6.475 1.86688 6.36562 1.92625 6.20937 1.92625C6.05312 1.92625 5.94375 1.86688 5.83125 1.79813C5.72187 1.73031 5.6125 1.64344 5.5 1.54281C5.27187 1.34156 5.03437 1.08531 4.80937 0.830312C4.49375 0.47 4.26562 0.179375 4.12812 0ZM3.85938 4.97L0 12.6919C0.00875 12.7263 0.025 12.7638 0.0675 12.8106C0.13 12.8825 0.241875 12.97 0.399063 13.0513C0.714063 13.22 1.20281 13.3731 1.79688 13.495C2.98438 13.7388 4.6 13.8638 6.20937 13.8638C7.81875 13.8638 9.43438 13.7388 10.6219 13.495C11.2156 13.3731 11.7063 13.22 12.0188 13.0513C12.1781 12.97 12.2875 12.8825 12.35 12.8106C12.3938 12.7638 12.4094 12.7263 12.4188 12.6919L8.55938 4.97C6.74688 5.40437 5.67188 5.40437 3.85938 4.97Z" fill={isActive ? "#FFFFFF" : "#596373"}/>
-</svg>
-  ),
+	// '전체' 아이콘
+	AllIcon: (isActive: boolean) => (
+		<svg
+			xmlns="http://www.w3.org/2000/svg"
+			width="14"
+			height="12"
+			viewBox="0 0 12 12"
+			fill="none"
+		>
+			<path
+				d="M8 12C7.63333 12 7.31956 11.8696 7.05867 11.6087C6.79778 11.3478 6.66711 11.0338 6.66667 10.6667V8C6.66667 7.63333 6.79733 7.31956 7.05867 7.05867C7.32 6.79778 7.63378 6.66711 8 6.66667H10.6667C11.0333 6.66667 11.3473 6.79733 11.6087 7.05867C11.87 7.32 12.0004 7.63378 12 8V10.6667C12 11.0333 11.8696 11.3473 11.6087 11.6087C11.3478 11.87 11.0338 12.0004 10.6667 12H8ZM8 5.33333C7.63333 5.33333 7.31956 5.20289 7.05867 4.942C6.79778 4.68111 6.66711 4.36711 6.66667 4V1.33333C6.66667 0.966667 6.79733 0.652889 7.05867 0.392C7.32 0.131111 7.63378 0.000444444 8 0H10.6667C11.0333 0 11.3473 0.130667 11.6087 0.392C11.87 0.653333 12.0004 0.967111 12 1.33333V4C12 4.36667 11.8696 4.68067 11.6087 4.942C11.3478 5.20333 11.0338 5.33378 10.6667 5.33333H8ZM1.33333 5.33333C0.966667 5.33333 0.652889 5.20289 0.392 4.942C0.131111 4.68111 0.000444444 4.36711 0 4V1.33333C0 0.966667 0.130667 0.652889 0.392 0.392C0.653333 0.131111 0.967111 0.000444444 1.33333 0H4C4.36667 0 4.68067 0.130667 4.942 0.392C5.20333 0.653333 5.33378 0.967111 5.33333 1.33333V4C5.33333 4.36667 5.20289 4.68067 4.942 4.942C4.68111 5.20333 4.36711 5.33378 4 5.33333H1.33333ZM1.33333 12C0.966667 12 0.652889 11.8696 0.392 11.6087C0.131111 11.3478 0.000444444 11.0338 0 10.6667V8C0 7.63333 0.130667 7.31956 0.392 7.05867C0.653333 6.79778 0.967111 6.66711 1.33333 6.66667H4C4.36667 6.66667 4.68067 6.79733 4.942 7.05867C5.20333 7.32 5.33378 7.63378 5.33333 8V10.6667C5.33333 11.0333 5.20289 11.3473 4.942 11.6087C4.68111 11.87 4.36711 12.0004 4 12H1.33333Z"
+				fill={isActive ? '#FFFFFF' : '#596373'}
+			/>
+		</svg>
+	),
+	// '상의' 아이콘
+	TopIcon: (isActive: boolean) => (
+		<svg
+			xmlns="http://www.w3.org/2000/svg"
+			width="14"
+			height="12"
+			viewBox="0 0 14 12"
+			fill="none"
+		>
+			<path
+				d="M8.58867 0.0048273L8.652 0.000160658L8.72667 0.00282729L8.802 0.0141606L8.87733 0.0341607L12.8773 1.36749C12.9976 1.40758 13.1039 1.48111 13.1838 1.57947C13.2638 1.67783 13.314 1.79693 13.3287 1.92283L13.3333 2.00016V5.33349C13.3333 5.49678 13.2734 5.65439 13.1649 5.77641C13.0563 5.89843 12.9068 5.97639 12.7447 5.99549L12.6667 6.00016H11.3333V10.6668C11.3334 11.0032 11.2064 11.3272 10.9777 11.5739C10.7489 11.8205 10.4354 11.9716 10.1 11.9968L10 12.0002H3.33333C2.99695 12.0003 2.67296 11.8732 2.4263 11.6445C2.17965 11.4158 2.02856 11.1023 2.00333 10.7668L2 10.6668V6.00016H0.666667C0.503378 6.00014 0.345775 5.94019 0.223752 5.83168C0.101729 5.72318 0.0237715 5.57366 0.00466668 5.41149L3.80479e-08 5.33349V2.00016C-4.27825e-05 1.87339 0.0360591 1.74924 0.104072 1.64226C0.172085 1.53528 0.26919 1.44991 0.384 1.39616L0.456 1.36749L4.456 0.0341607C4.55622 0.000780185 4.66294 -0.00830679 4.76736 0.00764891C4.87179 0.0236046 4.97092 0.0641461 5.0566 0.125932C5.14228 0.187719 5.21206 0.268981 5.26017 0.363022C5.30828 0.457064 5.33336 0.561193 5.33333 0.666827C5.33178 1.01282 5.46479 1.34586 5.70425 1.5956C5.94371 1.84533 6.27088 1.99221 6.61663 2.00519C6.96237 2.01817 7.29963 1.89624 7.55714 1.66517C7.81466 1.43409 7.97227 1.11196 7.99667 0.766827L8.00267 0.606827L8.014 0.531494L8.03867 0.442161L8.068 0.373494L8.10133 0.312161L8.14667 0.250161L8.19267 0.196827C8.22956 0.161272 8.26844 0.130161 8.30933 0.103494L8.37333 0.0681607L8.442 0.0388274L8.514 0.017494L8.58867 0.0048273Z"
+				fill={isActive ? '#FFFFFF' : '#596373'}
+			/>
+		</svg>
+	),
+	BottomIcon: (isActive: boolean) => (
+		<svg
+			xmlns="http://www.w3.org/2000/svg"
+			width="11"
+			height="15"
+			viewBox="0 0 11 15"
+			fill="none"
+		>
+			<path
+				d="M5.16506 9.62867e-05C3.77269 0.00594004 2.43881 0.230596 1.46 0.656346C0.9545 5.26441 0.43525 9.99469 0 13.961L3.8545 14.7365L4.92191 6.5226H5.56659L6.63691 14.7648H6.64372L10.6769 13.9435C10.186 9.51497 9.70747 5.085 9.21681 0.656346C8.00753 0.209252 6.5575 -0.00521621 5.16509 9.62867e-05H5.16506Z"
+				fill={isActive ? '#FFFFFF' : '#596373'}
+			/>
+		</svg>
+	),
+	OuterIcon: (isActive: boolean) => (
+		<svg
+			xmlns="http://www.w3.org/2000/svg"
+			width="15"
+			height="14"
+			viewBox="0 0 15 14"
+			fill="none"
+		>
+			<path
+				fillRule="evenodd"
+				clipRule="evenodd"
+				d="M8.49637 0C8.84578 2.37364e-05 9.19038 0.0814238 9.50287 0.23775L11.0254 0.99975C11.4829 1.22851 11.8888 1.54846 12.2181 1.93991C12.5474 2.33137 12.7931 2.78605 12.9401 3.276L14.6366 8.93025C14.7382 9.26864 14.7171 9.63203 14.577 9.95638C14.437 10.2807 14.1869 10.5452 13.8709 10.7032L13.1119 11.082C12.9648 11.1556 12.7976 11.1787 12.6361 11.148C12.4746 11.1172 12.3276 11.0342 12.2179 10.9117L11.8504 10.5L11.6494 11.814C11.5243 12.2969 11.2424 12.7245 10.8479 13.0297C10.4534 13.3348 9.96865 13.5003 9.46987 13.5H5.23087C4.73209 13.5003 4.24735 13.3348 3.85284 13.0297C3.45832 12.7245 3.1764 12.2969 3.05137 11.814L2.85037 10.5L2.48287 10.9117C2.37308 11.0344 2.22599 11.1175 2.0643 11.1482C1.9026 11.179 1.73528 11.1557 1.58812 11.082L0.829121 10.7032C0.513116 10.5452 0.263032 10.2807 0.122946 9.95638C-0.0171411 9.63203 -0.0382387 9.26864 0.0633712 8.93025L1.75987 3.276C1.9069 2.78605 2.15262 2.33137 2.48191 1.93991C2.81119 1.54846 3.21709 1.22851 3.67462 0.99975L5.19787 0.23775C5.51036 0.0814238 5.85496 2.37364e-05 6.20437 0H8.49637ZM8.49637 1.5H6.20437C6.1174 1.5001 6.03112 1.51533 5.94937 1.545L5.86912 1.5795L5.14912 1.9395L7.35037 5.3625L9.55162 1.938L8.83162 1.578C8.75376 1.53929 8.66976 1.51442 8.58337 1.5045L8.49637 1.5Z"
+				fill={isActive ? '#FFFFFF' : '#596373'}
+			/>
+		</svg>
+	),
+	DressIcon: (isActive: boolean) => (
+		<svg
+			xmlns="http://www.w3.org/2000/svg"
+			width="13"
+			height="14"
+			viewBox="0 0 13 14"
+			fill="none"
+		>
+			<path
+				d="M4.12812 0L3.49375 0.318125C3.49375 1.31906 3.51875 3.06375 3.92812 4.41062C5.75 4.85437 6.66875 4.85437 8.49063 4.41062C8.90625 3.06687 8.925 1.70625 8.92813 0.318125L8.29063 0C8.15313 0.179375 7.925 0.47 7.60937 0.830312C7.38437 1.08531 7.14688 1.34156 6.91875 1.54281C6.80625 1.64344 6.69688 1.73031 6.5875 1.79813C6.475 1.86688 6.36562 1.92625 6.20937 1.92625C6.05312 1.92625 5.94375 1.86688 5.83125 1.79813C5.72187 1.73031 5.6125 1.64344 5.5 1.54281C5.27187 1.34156 5.03437 1.08531 4.80937 0.830312C4.49375 0.47 4.26562 0.179375 4.12812 0ZM3.85938 4.97L0 12.6919C0.00875 12.7263 0.025 12.7638 0.0675 12.8106C0.13 12.8825 0.241875 12.97 0.399063 13.0513C0.714063 13.22 1.20281 13.3731 1.79688 13.495C2.98438 13.7388 4.6 13.8638 6.20937 13.8638C7.81875 13.8638 9.43438 13.7388 10.6219 13.495C11.2156 13.3731 11.7063 13.22 12.0188 13.0513C12.1781 12.97 12.2875 12.8825 12.35 12.8106C12.3938 12.7638 12.4094 12.7263 12.4188 12.6919L8.55938 4.97C6.74688 5.40437 5.67188 5.40437 3.85938 4.97Z"
+				fill={isActive ? '#FFFFFF' : '#596373'}
+			/>
+		</svg>
+	),
 
-  HomeIcon: (props: any) => (
-    <svg xmlns="http://www.w3.org/2000/svg" width="18" height="18" viewBox="0 0 18 18" fill="none" {...props}>
-    <path
-      d="M16.75 15.5V7C16.75 6.84475 16.7139 6.69164 16.6444 6.55279C16.575 6.41393 16.4742 6.29315 16.35 6.2L9.35 0.95C9.1769 0.820178 8.96637 0.75 8.75 0.75C8.53363 0.75 8.3231 0.820178 8.15 0.95L1.15 6.2C1.0258 6.29315 0.925 6.41393 0.855573 6.55279C0.786145 6.69164 0.75 6.84475 0.75 7V15.5C0.75 15.7652 0.855357 16.0196 1.04289 16.2071C1.23043 16.3946 1.48478 16.5 1.75 16.5H5.75C6.01522 16.5 6.26957 16.3946 6.45711 16.2071C6.64464 16.0196 6.75 15.7652 6.75 15.5V12.5C6.75 12.2348 6.85536 11.9804 7.04289 11.7929C7.23043 11.6054 7.48478 11.5 7.75 11.5H9.75C10.0152 11.5 10.2696 11.6054 10.4571 11.7929C10.6446 11.9804 10.75 12.2348 10.75 12.5V15.5C10.75 15.7652 10.8554 16.0196 11.0429 16.2071C11.2304 16.3946 11.4848 16.5 11.75 16.5H15.75C16.0152 16.5 16.2696 16.3946 16.4571 16.2071C16.6446 16.0196 16.75 15.7652 16.75 15.5Z"
-      fill="currentColor"
-      stroke="currentColor"
-      strokeWidth="1.5"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-  </svg>
-  ),
-  ProductIcon: (props: any) =>(
-    <svg xmlns="http://www.w3.org/2000/svg" width="20" height="18" viewBox="0 0 20 18" fill="none" {...props}>
-    <path
-      d="M12.75 0.75L18.75 2.75V7.75H15.75V15.75C15.75 16.0152 15.6446 16.2696 15.4571 16.4571C15.2696 16.6446 15.0152 16.75 14.75 16.75H4.75C4.48478 16.75 4.23043 16.6446 4.04289 16.4571C3.85536 16.2696 3.75 16.0152 3.75 15.75V7.75H0.75V2.75L6.75 0.75C6.75 1.54565 7.06607 2.30871 7.62868 2.87132C8.19129 3.43393 8.95435 3.75 9.75 3.75C10.5456 3.75 11.3087 3.43393 11.8713 2.87132C12.4339 2.30871 12.75 1.54565 12.75 0.75Z"
-      fill="currentColor"
-      stroke="currentColor"
-      strokeWidth="1.5"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-  </svg>
-  ),
-  FittingIcon: (props: any) =>(
-    <svg xmlns="http://www.w3.org/2000/svg" width="20" height="16" viewBox="0 0 20 16" fill="none" {...props}>
-    <path
-      d="M9.9991 1.49972C9.3471 1.49972 8.7911 1.91572 8.5841 2.49972C8.55246 2.59409 8.50236 2.68121 8.43671 2.75601C8.37106 2.83082 8.29118 2.89181 8.20173 2.93543C8.11227 2.97905 8.01503 3.00443 7.91566 3.0101C7.8163 3.01576 7.7168 3.00158 7.62297 2.96841C7.52914 2.93523 7.44285 2.8837 7.36913 2.81684C7.29541 2.74997 7.23573 2.66911 7.19358 2.57895C7.15143 2.48879 7.12765 2.39114 7.12362 2.2917C7.11958 2.19225 7.13538 2.093 7.1701 1.99972C7.32693 1.55628 7.58583 1.15593 7.9259 0.830988C8.26597 0.506045 8.67768 0.265616 9.12779 0.129113C9.5779 -0.00739013 10.0538 -0.0361434 10.5171 0.0451743C10.9804 0.126492 11.418 0.315602 11.7947 0.597238C12.1715 0.878874 12.4767 1.24514 12.6858 1.66648C12.8948 2.08781 13.0019 2.5524 12.9983 3.02275C12.9948 3.49309 12.8807 3.956 12.6652 4.37412C12.4498 4.79223 12.139 5.15383 11.7581 5.42972C11.4571 5.64872 11.1901 5.86772 11.0011 6.09672C10.8151 6.32272 10.7491 6.50372 10.7491 6.65972C10.7492 6.7786 10.7814 6.89524 10.8423 6.99733C10.9032 7.09942 10.9905 7.18317 11.0951 7.23972L18.7021 11.3517C19.186 11.6135 19.5687 12.0292 19.7898 12.533C20.0108 13.0368 20.0576 13.6 19.9226 14.1333C19.7877 14.6667 19.4788 15.1398 19.0447 15.4778C18.6106 15.8159 18.0763 15.9995 17.5261 15.9997H2.4721C1.92194 15.9995 1.38755 15.8159 0.953493 15.4778C0.519435 15.1398 0.210493 14.6667 0.0755477 14.1333C-0.0593975 13.6 -0.0126392 13.0368 0.208424 12.533C0.429487 12.0292 0.812231 11.6135 1.2961 11.3517L7.3961 8.05372C7.48277 8.00691 7.57782 7.97762 7.67581 7.96754C7.7738 7.95745 7.87281 7.96677 7.9672 7.99495C8.06159 8.02313 8.14951 8.06963 8.22593 8.13179C8.30235 8.19395 8.36578 8.27055 8.4126 8.35722C8.45941 8.4439 8.4887 8.53894 8.49879 8.63693C8.50887 8.73492 8.49955 8.83394 8.47137 8.92833C8.44319 9.02272 8.39669 9.11064 8.33453 9.18706C8.27237 9.26348 8.19577 9.32691 8.1091 9.37372L2.0091 12.6707C1.81857 12.7737 1.66784 12.9373 1.58078 13.1356C1.49371 13.3339 1.4753 13.5556 1.52844 13.7655C1.58159 13.9754 1.70326 14.1617 1.87419 14.2947C2.04512 14.4276 2.25553 14.4998 2.4721 14.4997H17.5261C17.7427 14.4998 17.9531 14.4276 18.124 14.2947C18.2949 14.1617 18.4166 13.9754 18.4698 13.7655C18.5229 13.5556 18.5045 13.3339 18.4174 13.1356C18.3304 12.9373 18.1796 12.7737 17.9891 12.6707L10.3821 8.55872C10.0395 8.37357 9.75335 8.09924 9.5539 7.76478C9.35446 7.43031 9.24914 7.04814 9.2491 6.65872C9.2491 6.03572 9.5271 5.52772 9.8431 5.14372C10.1571 4.76372 10.5501 4.45372 10.8781 4.21572C11.1352 4.03001 11.3269 3.76755 11.4255 3.46608C11.5241 3.16461 11.5246 2.83964 11.427 2.53785C11.3293 2.23607 11.1385 1.97302 10.8819 1.78649C10.6254 1.59997 10.3163 1.49957 9.9991 1.49972Z"
-      fill="currentColor"
-    />
-  </svg>
-  ),
-  ClosetIcon: (props: any) =>(
-    <svg xmlns="http://www.w3.org/2000/svg" width="15" height="18" viewBox="0 0 15 18" fill="none" {...props}>
-    <path
-      d="M7.45581 0.75V16.6001M10.5039 8.06543V9.28467M4.40771 8.06543V9.28467M12.9424 0.75H1.96924C1.64588 0.75 1.33576 0.878455 1.10711 1.10711C0.878455 1.33576 0.75 1.64588 0.75 1.96924V15.3809C0.75 15.7042 0.878455 16.0143 1.10711 16.243C1.33576 16.4716 1.64588 16.6001 1.96924 16.6001H12.9424C13.2657 16.6001 13.5759 16.4716 13.8045 16.243C14.0332 16.0143 14.1616 15.7042 14.1616 15.3809V1.96924C14.1616 1.64588 14.0332 1.33576 13.8045 1.10711C13.5759 0.878455 13.2657 0.75 12.9424 0.75Z"
-      stroke="currentColor"
-      strokeWidth="1.5"
-      strokeLinecap="round"
-      strokeLinejoin="round"
-    />
-  </svg>
-  ),
+	HomeIcon: (props: any) => (
+		<svg
+			xmlns="http://www.w3.org/2000/svg"
+			width="18"
+			height="18"
+			viewBox="0 0 18 18"
+			fill="none"
+			{...props}
+		>
+			<path
+				d="M16.75 15.5V7C16.75 6.84475 16.7139 6.69164 16.6444 6.55279C16.575 6.41393 16.4742 6.29315 16.35 6.2L9.35 0.95C9.1769 0.820178 8.96637 0.75 8.75 0.75C8.53363 0.75 8.3231 0.820178 8.15 0.95L1.15 6.2C1.0258 6.29315 0.925 6.41393 0.855573 6.55279C0.786145 6.69164 0.75 6.84475 0.75 7V15.5C0.75 15.7652 0.855357 16.0196 1.04289 16.2071C1.23043 16.3946 1.48478 16.5 1.75 16.5H5.75C6.01522 16.5 6.26957 16.3946 6.45711 16.2071C6.64464 16.0196 6.75 15.7652 6.75 15.5V12.5C6.75 12.2348 6.85536 11.9804 7.04289 11.7929C7.23043 11.6054 7.48478 11.5 7.75 11.5H9.75C10.0152 11.5 10.2696 11.6054 10.4571 11.7929C10.6446 11.9804 10.75 12.2348 10.75 12.5V15.5C10.75 15.7652 10.8554 16.0196 11.0429 16.2071C11.2304 16.3946 11.4848 16.5 11.75 16.5H15.75C16.0152 16.5 16.2696 16.3946 16.4571 16.2071C16.6446 16.0196 16.75 15.7652 16.75 15.5Z"
+				fill="currentColor"
+				stroke="currentColor"
+				strokeWidth="1.5"
+				strokeLinecap="round"
+				strokeLinejoin="round"
+			/>
+		</svg>
+	),
+	ProductIcon: (props: any) =>(
+		<svg
+			xmlns="http://www.w3.org/2000/svg"
+			width="20"
+			height="18"
+			viewBox="0 0 20 18"
+			fill="none"
+			{...props}
+		>
+			<path
+				d="M12.75 0.75L18.75 2.75V7.75H15.75V15.75C15.75 16.0152 15.6446 16.2696 15.4571 16.4571C15.2696 16.6446 15.0152 16.75 14.75 16.75H4.75C4.48478 16.75 4.23043 16.6446 4.04289 16.4571C3.85536 16.2696 3.75 16.0152 3.75 15.75V7.75H0.75V2.75L6.75 0.75C6.75 1.54565 7.06607 2.30871 7.62868 2.87132C8.19129 3.43393 8.95435 3.75 9.75 3.75C10.5456 3.75 11.3087 3.43393 11.8713 2.87132C12.4339 2.30871 12.75 1.54565 12.75 0.75Z"
+				fill="currentColor"
+				stroke="currentColor"
+				strokeWidth="1.5"
+				strokeLinecap="round"
+				strokeLinejoin="round"
+			/>
+		</svg>
+	),
+	FittingIcon: (props: any) =>(
+		<svg
+			xmlns="http://www.w3.org/2000/svg"
+			width="20"
+			height="16"
+			viewBox="0 0 20 16"
+			fill="none"
+			{...props}
+		>
+			<path
+				d="M9.9991 1.49972C9.3471 1.49972 8.7911 1.91572 8.5841 2.49972C8.55246 2.59409 8.50236 2.68121 8.43671 2.75601C8.37106 2.83082 8.29118 2.89181 8.20173 2.93543C8.11227 2.97905 8.01503 3.00443 7.91566 3.0101C7.8163 3.01576 7.7168 3.00158 7.62297 2.96841C7.52914 2.93523 7.44285 2.8837 7.36913 2.81684C7.29541 2.74997 7.23573 2.66911 7.19358 2.57895C7.15143 2.48879 7.12765 2.39114 7.12362 2.2917C7.11958 2.19225 7.13538 2.093 7.1701 1.99972C7.32693 1.55628 7.58583 1.15593 7.9259 0.830988C8.26597 0.506045 8.67768 0.265616 9.12779 0.129113C9.5779 -0.00739013 10.0538 -0.0361434 10.5171 0.0451743C10.9804 0.126492 11.418 0.315602 11.7947 0.597238C12.1715 0.878874 12.4767 1.24514 12.6858 1.66648C12.8948 2.08781 13.0019 2.5524 12.9983 3.02275C12.9948 3.49309 12.8807 3.956 12.6652 4.37412C12.4498 4.79223 12.139 5.15383 11.7581 5.42972C11.4571 5.64872 11.1901 5.86772 11.0011 6.09672C10.8151 6.32272 10.7491 6.50372 10.7491 6.65972C10.7492 6.7786 10.7814 6.89524 10.8423 6.99733C10.9032 7.09942 10.9905 7.18317 11.0951 7.23972L18.7021 11.3517C19.186 11.6135 19.5687 12.0292 19.7898 12.533C20.0108 13.0368 20.0576 13.6 19.9226 14.1333C19.7877 14.6667 19.4788 15.1398 19.0447 15.4778C18.6106 15.8159 18.0763 15.9995 17.5261 15.9997H2.4721C1.92194 15.9995 1.38755 15.8159 0.953493 15.4778C0.519435 15.1398 0.210493 14.6667 0.0755477 14.1333C-0.0593975 13.6 -0.0126392 13.0368 0.208424 12.533C0.429487 12.0292 0.812231 11.6135 1.2961 11.3517L7.3961 8.05372C7.48277 8.00691 7.57782 7.97762 7.67581 7.96754C7.7738 7.95745 7.87281 7.96677 7.9672 7.99495C8.06159 8.02313 8.14951 8.06963 8.22593 8.13179C8.30235 8.19395 8.36578 8.27055 8.4126 8.35722C8.45941 8.4439 8.4887 8.53894 8.49879 8.63693C8.50887 8.73492 8.49955 8.83394 8.47137 8.92833C8.44319 9.02272 8.39669 9.11064 8.33453 9.18706C8.27237 9.26348 8.19577 9.32691 8.1091 9.37372L2.0091 12.6707C1.81857 12.7737 1.66784 12.9373 1.58078 13.1356C1.49371 13.3339 1.4753 13.5556 1.52844 13.7655C1.58159 13.9754 1.70326 14.1617 1.87419 14.2947C2.04512 14.4276 2.25553 14.4998 2.4721 14.4997H17.5261C17.7427 14.4998 17.9531 14.4276 18.124 14.2947C18.2949 14.1617 18.4166 13.9754 18.4698 13.7655C18.5229 13.5556 18.5045 13.3339 18.4174 13.1356C18.3304 12.9373 18.1796 12.7737 17.9891 12.6707L10.3821 8.55872C10.0395 8.37357 9.75335 8.09924 9.5539 7.76478C9.35446 7.43031 9.24914 7.04814 9.2491 6.65872C9.2491 6.03572 9.5271 5.52772 9.8431 5.14372C10.1571 4.76372 10.5501 4.45372 10.8781 4.21572C11.1352 4.03001 11.3269 3.76755 11.4255 3.46608C11.5241 3.16461 11.5246 2.83964 11.427 2.53785C11.3293 2.23607 11.1385 1.97302 10.8819 1.78649C10.6254 1.59997 10.3163 1.49957 9.9991 1.49972Z"
+				fill="currentColor"
+			/>
+		</svg>
+	),
+	ClosetIcon: (props: any) =>(
+		<svg
+			xmlns="http://www.w3.org/2000/svg"
+			width="15"
+			height="18"
+			viewBox="0 0 15 18"
+			fill="none"
+			{...props}
+		>
+			<path
+				d="M7.45581 0.75V16.6001M10.5039 8.06543V9.28467M4.40771 8.06543V9.28467M12.9424 0.75H1.96924C1.64588 0.75 1.33576 0.878455 1.10711 1.10711C0.878455 1.33576 0.75 1.64588 0.75 1.96924V15.3809C0.75 15.7042 0.878455 16.0143 1.10711 16.243C1.33576 16.4716 1.64588 16.6001 1.96924 16.6001H12.9424C13.2657 16.6001 13.5759 16.4716 13.8045 16.243C14.0332 16.0143 14.1616 15.7042 14.1616 15.3809V1.96924C14.1616 1.64588 14.0332 1.33576 13.8045 1.10711C13.5759 0.878455 13.2657 0.75 12.9424 0.75Z"
+				stroke="currentColor"
+				strokeWidth="1.5"
+				strokeLinecap="round"
+				strokeLinejoin="round"
+			/>
+		</svg>
+	),
 
 
 };
