+++ src/app/App.tsx
@@ -3,9 +3,9 @@ import { router } from './router/router';
 
 function App() {
 
-  return (
-    <RouterProvider router={router} />
-  );
+	return (
+		<RouterProvider router={router} />
+	);
 }
 
 export default App;
