+++ src/app/main.tsx
@@ -1,9 +1,9 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import './index.css'
-import './globals.css'
+import App from './App'
+import '@/src/shared/styles/globals.css'
+import '@/src/shared/styles/index.css'
 
-import App from './App.tsx'
 
 createRoot(document.getElementById('root')!).render(
 	<StrictMode>
