+++ src/pages/home/home-page.tsx
@@ -1,0 +1,7 @@
+import HomePageWidget from '@/src/widgets/home/HomePageWidget';
+
+const HomePage = () => {
+	return <HomePageWidget />;
+};
+
+export default HomePage;
