+++ src/shared/components/common/CategoryBar.tsx
