+++ src/shared/configs/path.ts
